refactor: move login request and response DTOs into auth folder
