+++ drumigo/src/main/java/com/ftn/drumigo/dto/auth/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.ftn.drumigo.dto;
+package com.ftn.drumigo.dto.auth.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
